hv: fix the race issue during vm create/destroy and complete_ioreq
If the ioreq polling mode is enabled, it's possible to have race
issues in two places:

1. VM destroy phase
     CPU1		      CPU2
                              default_idle
                                handle_complete_ioreq
                                  vcpu = get_ever_run_vcpu(pcpu_id)
 offline_vcpu()                   vm = vcpu->vm
   ever_run_vcpu = NULL           ....
 further_cleanup                  vm access trigger crash

2. VM create phase
     CPU1		      CPU2
			      default_idle
   create_vcpu		        handle_complete_ioreq
     ever_run_vcpu = vcpu         vcpu = get_ever_run_vcpu(pcpu_id)
                                  vm = vcpu->vm
                                  ...
                                  vm access trigger crash

     vcpu->vm = vm
     other vcpu fields init

The fixing is like
For case 1, we make sure the io_polling flag disable polling before
we pause vm in shutdown_vm. That will prevent the handle_complete_ioreq
accessing further vcpu/vm resource. And avoid the race issue.

For case 2, we only assign ever_run_vcpu when vcpu is scheduled to target
pcpu. The vm access is fine at that time and avoid the race issue.

Signed-off-by: Yin Fengwei <fengwei.yin@intel.com>